update acorrding to asaf comments
