+++ GCM/App/register/RegisterModel.java
@@ -63,15 +63,4 @@ public class RegisterModel {
 
 	public void back() { gcmClient.back(); }
 	
-	public void showAlert(Alert.AlertType alertType, Window owner, String title, String message) {
-        Alert alert = new Alert(alertType);
-        alert.setTitle(title);
-        alert.setHeaderText(null);
-        alert.setContentText(message);
-        alert.initOwner(owner);
-        alert.show();
-    }
-	
-
-	
 }
