+++ GCM/Database/database/tests/DatabaseExecutorTest.java
@@ -1,4 +1,4 @@
-package database.Tests;
+package database.tests;
 
 import static org.junit.Assert.fail;
